requirements.txt
pytest==8.3.2
aiohttp==3.8.3
aiogram==2.25.2
asyncpg==0.29.0
aioresponses==0.7.6
python-dotenv==1.0.1
pytest-asyncio==0.23.8